PB-1515 : Refactor Asset & CollectionAsset views with a mixin
